refactor: replace old select, define customStyles in one place
