+++ keep-ui/app/(keep)/alerts/alert-push-alert-to-server-modal.tsx
@@ -10,9 +10,9 @@ import Modal from "@/components/ui/Modal";
 import { useProviders } from "utils/hooks/useProviders";
 import ImageWithFallback from "@/components/ImageWithFallback";
 import { useAlerts } from "utils/hooks/useAlerts";
-import Select from "@/components/ui/Select";
 import { useApi } from "@/shared/lib/hooks/useApi";
 import { KeepApiError } from "@/shared/api";
+import { Select } from "@/shared/ui";
 
 import { useRevalidateMultiple } from "@/shared/lib/state-utils";
