+++ keep-ui/entities/incidents/ui/IncidentIconName/index.ts
@@ -1,0 +1,1 @@
+export { IncidentIconName } from "./IncidentIconName";
