+++ keep-ui/features/same-incidents-in-the-past/ui/change-same-incident-in-the-past-form.tsx
@@ -1,5 +1,4 @@
 import { Button, Divider, Title } from "@tremor/react";
-import Select from "@/components/ui/Select";
 import { useRouter } from "next/navigation";
 import { FormEvent, useState } from "react";
 import { useIncidents, usePollIncidents } from "@/utils/hooks/useIncidents";
@@ -7,6 +6,7 @@ import Loading from "@/app/(keep)/loading";
 import type { IncidentDto } from "@/entities/incidents/model";
 import { useIncidentActions } from "@/entities/incidents/model";
 import { getIncidentName } from "@/entities/incidents/lib/utils";
+import { Select } from "@/shared/ui";
 
 interface ChangeSameIncidentInThePastFormProps {
   incident: IncidentDto;
