+++ keep-ui/shared/ui/VerticalRoundedList/index.ts
@@ -1,0 +1,1 @@
+export { VerticalRoundedList } from "./VerticalRoundedList";
