+++ keep-ui/shared/ui/VerticalRoundedList/vertical-rounded-list.css
