+++ keep-ui/shared/ui/index.ts
@@ -11,6 +11,8 @@ export { SeverityLabel } from "./SeverityLabel";
 export { DropdownMenu } from "./DropdownMenu";
 export { SeverityBorderIcon } from "./SeverityBorderIcon";
 export { TableSeverityCell } from "./TableSeverityCell";
+export { Select } from "./Select";
+export { VerticalRoundedList } from "./VerticalRoundedList";
 
 export { getCommonPinningStylesAndClassNames } from "./utils/table-utils";
 export { ThemeScript, WatchUpdateTheme, ThemeControl } from "./theme";
